Update types to bigint.

+++ schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    tg_id       int4 PRIMARY KEY,
+    tg_id       bigint PRIMARY KEY,
     first_name  text,
     last_name   text,
     username    text,
@@ -10,23 +10,23 @@ CREATE TABLE users (
 -- VALUES (59352582, 'Вячеслав', TRUE);
 
 CREATE TABLE chats (
-    tg_id        int4 PRIMARY KEY,
+    tg_id        bigint PRIMARY KEY,
     title        text,
     chain_period int DEFAULT 30
 );
 
 CREATE TABLE updates (
     tg_id   int4 PRIMARY KEY,
-    chat_id int4,
+    chat_id bigint,
     created timestamptz DEFAULT now(),
     dump    bytea
 );
 
 CREATE TABLE messages (
-    chat_id   int4 REFERENCES chats(tg_id),
+    chat_id   int4   REFERENCES chats(tg_id),
     tg_id     int4,
-    user_id   int4 REFERENCES users(tg_id),
-    update_id int4 REFERENCES updates(tg_id),
+    user_id   bigint REFERENCES users(tg_id),
+    update_id int4   REFERENCES updates(tg_id),
 
     text      text,
     date      timestamptz,
@@ -36,7 +36,7 @@ CREATE TABLE messages (
 
 CREATE TABLE reminders (
     id       serial4     PRIMARY KEY,
-    chat_id  int4        REFERENCES chats(tg_id),
+    chat_id  bigint      REFERENCES chats(tg_id),
     datetime timestamptz NOT NULL,
     text     text        NOT NULL,
     isactive bool        NOT NULL
